Initial 3 songs
Each song is split into 20-50 short song clips, and each clip has a directory containing the voice version of the clip, a karaoke backing (no vocals) version of the same clip, and a lyric.txt file with the lyric sung in the clip.

Generally the lyric is 1-2 lines of the song. Some directories contain filler video for use in instrumental parts of the song.

Each song directory has a "songinfo.json" file with basic info about the song, and a full version of the karaoke backing vocals for audio mixing before playback.